+++ app/routes/home.tsx
@@ -1,9 +1,9 @@
 import type { Route } from "./+types/home";
 import Hero from "~/components/Hero";
 import Mission from "~/components/Mission";
-import Navbar from "~/components/Navbar";
-import Socials from "~/components/Socials";
-import Team from "~/components/Team";
+import Navbar from "~/components/nav/NavBar";
+import Socials from "~/components/socials/Socials";
+import Team from "~/components/team/Team";
 
 export function meta({ }: Route.MetaArgs) {
     return [
